requirements.txt
adjustText==0.7.3
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
requests==2.27.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
tabulate==0.8.10
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
